introduce BaseBuilderBlock
